# docs/swagger.yaml
basePath: /
definitions:
...
  user.SignInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.SignUpCompany:
    properties:
      company_account:
        type: string
      email:
        type: string
      location:
        type: string
      manager_phone_number:
        type: string
      name:
        type: string
      password:
        type: string
      passwordConfirm:
        type: string
      phone_number:
        type: string
      photo:
        $ref: '#/definitions/utils.Base64Struct'
      prole_id:
        type: string
      role_id:
        type: string
    required:
    - email
    - name
    - password
    - passwordConfirm
    - phone_number
    type: object
  user.SignUpInfluencer:
    properties:
      audience_interests:
        type: string
      average_likes:
        type: number
      avg_comments:
        type: number
      avg_likes:
        type: number
      avg_reel_plays:
        type: number
      avg_views:
        type: number
      bio:
        type: string
      email:
        type: string
      engagement_rate:
        type: number
      followers:
        type: number
      gender_split:
        type: string
      ig_name:
        type: string
      location:
        type: string
      name:
        type: string
      password:
        type: string
      passwordConfirm:
        type: string
      phone_number:
        type: string
      photo:
        $ref: '#/definitions/utils.Base64Struct'
      popular_posts:
        type: string
      role_id:
        type: string
      total_posts:
        type: number
    required:
    - email
    - ig_name
    - name
    - password
    - passwordConfirm
    type: object
  user.SignUpInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      passwordConfirm:
        minLength: 8
        type: string
      photo:
        $ref: '#/definitions/utils.Base64Struct'
    required:
    - email
    - name
    - password
    - passwordConfirm
    type: object
...
  utils.Base64Struct:
    properties:
      base64:
        type: string
    type: object
...
host: http://103.168.56.249:8080
info:
  contact: {}
  description: This is a sample API with Fiber and Swagger
  title: Ramen API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Sign in user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Sign in user
      tags:
      - Auth
  /auth/signup/admin:
    post:
      consumes:
      - application/json
      description: Create a new admin
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.SignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a new admin
      tags:
      - Auth
  /auth/signup/company:
    post:
      consumes:
      - application/json
      description: Sign up company
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.SignUpCompany'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Sign up company
      tags:
      - Auth
  /auth/signup/influencer:
    post:
      consumes:
      - application/json
      description: Sign up influencer
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.SignUpInfluencer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Sign up influencer
      tags:
      - Auth
...
swagger: "2.0"